+++ .eslintrc.js
@@ -1,37 +1,3 @@
 module.exports = {
-    "env": {
-        "browser": true,
-        "commonjs": true,
-        "node": true,
-        "es6": true,
-        "jasmine": true
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-        "ecmaVersion": 8
-    },
-    "rules": {
-        "indent": [
-            "error",
-            4
-        ],
-        "linebreak-style": [
-            "error",
-            "unix"
-        ],
-        "quotes": [
-            "error",
-            "single",
-            {
-                "allowTemplateLiterals": true
-            }
-        ],
-        "semi": [
-            "error",
-            "always"
-        ]
-    },
-    "globals": {
-        "Mzp": true,
-    }
+    extends: "@mozilla-protocol/eslint-config",
 };