+++ gulpfile.js/tasks/compress-css.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const gulp = require('gulp');
 const cleanCSS = require('gulp-clean-css');
 const rename = require('gulp-rename');