+++ gulpfile.js/tasks/compress-js.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const gulp = require('gulp');
 const uglify = require('gulp-uglify');
 const plumber = require('gulp-plumber');