+++ gulpfile.js/tasks/concat-js.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const gulp = require('gulp');
 const concat = require('gulp-concat');
 const plumber = require('gulp-plumber');