+++ gulpfile.js/tasks/copy-sass.js
@@ -1,3 +1,5 @@
+'use strict';
+
 const gulp = require('gulp');
 const config = require('../config').sassCopy;
 const merge = require('merge-stream');