+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+* **js:** Update ESLint to consume @mozilla-protocol/eslint-config (##85)
 * **css:** Rename download button theme to product button theme (#273)
 * **css:** Navigation items that are links without menus get hover styles (#304)
 * **css:** Let menu item titles not be links (#307)